Move internal/pkg/core/server to pkg/core/server
This package is used by the Fleet Server and is not available anymore
in the beats repository.
